Allow other Go packages to import/use the miniooni CLI
Part of https://github.com/ooni/probe-engine/issues/547.
